feat: remove redundancy from group
